Add files via upload
My attempt at question 1 of the 2021 BIO.  I have tried to do it without recursion as I am trying to think about it in the way a student would who has not come across recursion yet i.e. starting A'Level / IB Computer Science.  I have focused on working out whether one string of up to 6 characters entered is a pat and not the two strings as requested by the question.